# src/docs/swagger.yaml
openapi: 3.0.3
info:
    title: Fetcher API
    version: 1.0.0
    description: |
        REST API for discovering and executing fetchers and managing their persisted results.

        Protected operations accept JWT bearer tokens issued for this API. Depending on the
        operation, the token must represent either a user with the documented minimum system
        role or an authenticated service.
servers:
    - url: /
...
paths:
...
    /api/v1/fetchers/{fetcherId}/fetchResults/generate:
        parameters:
            - $ref: '#/components/parameters/fetcherId'
            - $ref: '#/components/parameters/isAsync'
        post:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Generate a fetch result
            operationId: generateFetchResult
            description: |
                Resolves and persists the result identified by `fetcherId`, `effectiveAt` and the
                canonical hash of `fetcherConfig`. Requests for that same logical key share one entry.

                Resolution matrix when no reusable result exists:

                | Fetcher capability | `CAPTURE` | `REPLAY` |
                | --- | --- | --- |
                | `SNAPSHOT` | Calls the external source and stores the observation. | Does not call the external source; creates `UNAVAILABLE` with `PAST_SNAPSHOT_NOT_CAPTURED`. |
                | `HISTORICAL` | Calls the external source for `effectiveAt`. | Calls the historical external source for `effectiveAt`. |

                Existing result behavior:

                - `COMPLETED`, `UNAVAILABLE` and `IN_PROGRESS` are returned as-is; no new external call is made.
                - `FAILED` is reclaimed using the same document and reset to `IN_PROGRESS`. A retried
                  `SNAPSHOT` in `REPLAY` becomes `UNAVAILABLE` without an external call.
                - The stored `acquisitionMode` describes the request that created or last retried the result.
                  A cache hit can therefore return a different mode from the current request.

                With `isAsync=false`, a new external operation finishes before responding. With
                `isAsync=true`, a new external operation returns `IN_PROGRESS` and continues in the
                background. Cache hits and `SNAPSHOT` + `REPLAY` can return a terminal result immediately
                even when `isAsync=true`.

                Accepts a service token or a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/GenerateFetchResultBody'
            responses:
                '200':
                    description: |
                        Result claimed, resolved, or reused. Domain outcomes such as `UNAVAILABLE` and
                        external-fetch `FAILED` are returned as successful HTTP responses; inspect `data.status`.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                            examples:
                                asynchronousCapture:
                                    summary: New asynchronous capture still being processed
                                    value:
                                        success: true
                                        message: Fetch result started asynchronously
                                        httpStatus: 200
                                        appCode: SUCCESS
                                        data:
                                            _id: 67c9ff7b2ddf1b3940f9f321
                                            startDate: '2026-04-10T09:30:00.000Z'
                                            endDate: null
                                            effectiveAt: '2026-04-10T09:30:00.000Z'
                                            temporalCapability: SNAPSHOT
                                            acquisitionMode: CAPTURE
                                            status: IN_PROGRESS
                                            unavailableReason: null
                                            fetcherConfig:
                                                owner: governify
                                                repository: fetcher
                                            configHash: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                                            data: null
                                            createdAt: '2026-04-10T09:30:00.000Z'
                                            updatedAt: '2026-04-10T09:30:00.000Z'
                                            __v: 0
                                        error: null
                                completedCapture:
                                    summary: Snapshot captured successfully
                                    value:
                                        success: true
                                        message: Fetch result generated
                                        httpStatus: 200
                                        appCode: SUCCESS
                                        data:
                                            _id: 67c9ff7b2ddf1b3940f9f322
                                            startDate: '2026-04-10T09:29:59.950Z'
                                            endDate: '2026-04-10T09:30:00.120Z'
                                            effectiveAt: '2026-04-10T09:30:00.000Z'
                                            temporalCapability: SNAPSHOT
                                            acquisitionMode: CAPTURE
                                            status: COMPLETED
                                            unavailableReason: null
                                            fetcherConfig:
                                                owner: governify
                                                repository: fetcher
                                            configHash: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                                            data:
                                                issues:
                                                    - number: 42
                                                      column: In progress
                                            createdAt: '2026-04-10T09:29:59.950Z'
                                            updatedAt: '2026-04-10T09:30:00.120Z'
                                            __v: 0
                                        error: null
                                unavailableSnapshotReplay:
                                    summary: Past snapshot was not captured
                                    value:
                                        success: true
                                        message: Fetch result generated
                                        httpStatus: 200
                                        appCode: SUCCESS
                                        data:
                                            _id: 67c9ff7b2ddf1b3940f9f323
                                            startDate: '2026-04-17T12:00:00.000Z'
                                            endDate: '2026-04-17T12:00:00.005Z'
                                            effectiveAt: '2026-04-10T09:30:00.000Z'
                                            temporalCapability: SNAPSHOT
                                            acquisitionMode: REPLAY
                                            status: UNAVAILABLE
                                            unavailableReason: PAST_SNAPSHOT_NOT_CAPTURED
                                            fetcherConfig:
                                                owner: governify
                                                repository: fetcher
                                            configHash: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                                            data: null
                                            createdAt: '2026-04-17T12:00:00.000Z'
                                            updatedAt: '2026-04-17T12:00:00.005Z'
                                            __v: 0
                                        error: null
                                completedHistoricalReplay:
                                    summary: Historical source queried during replay
                                    value:
                                        success: true
                                        message: Fetch result generated
                                        httpStatus: 200
                                        appCode: SUCCESS
                                        data:
                                            _id: 67c9ff7b2ddf1b3940f9f324
                                            startDate: '2026-04-17T12:00:00.000Z'
                                            endDate: '2026-04-17T12:00:00.220Z'
                                            effectiveAt: '2026-04-10T09:30:00.000Z'
                                            temporalCapability: HISTORICAL
                                            acquisitionMode: REPLAY
                                            status: COMPLETED
                                            unavailableReason: null
                                            fetcherConfig:
                                                latitude: 37.39
                                                longitude: -5.99
                                            configHash: bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
                                            data:
                                                temperature: 21.4
                                                unit: celsius
                                            createdAt: '2026-04-17T12:00:00.000Z'
                                            updatedAt: '2026-04-17T12:00:00.220Z'
                                            __v: 0
                                        error: null
                                failedExternalFetch:
                                    summary: External source failed; the next request retries this document
                                    value:
                                        success: true
                                        message: Fetch result generated
                                        httpStatus: 200
                                        appCode: SUCCESS
                                        data:
                                            _id: 67c9ff7b2ddf1b3940f9f325
                                            startDate: '2026-04-10T09:30:00.000Z'
                                            endDate: '2026-04-10T09:30:03.000Z'
                                            effectiveAt: '2026-04-10T09:30:00.000Z'
                                            temporalCapability: SNAPSHOT
                                            acquisitionMode: CAPTURE
                                            status: FAILED
                                            unavailableReason: null
                                            fetcherConfig:
                                                owner: governify
                                                repository: fetcher
                                            configHash: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                                            data: null
                                            createdAt: '2026-04-10T09:29:55.000Z'
                                            updatedAt: '2026-04-10T09:30:03.000Z'
                                            __v: 0
                                        error: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/fetchers/{fetcherId}/fetchResults:
        parameters:
            - $ref: '#/components/parameters/fetcherId'
        get:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Get fetch results by fetcher ID
            operationId: getFetchResultsByFetcherId
            description: |
                Retrieves all fetch results associated with a given fetcher.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetch results retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Delete fetch results by fetcher ID
            operationId: deleteFetchResultsByFetcherId
            description: |
                Deletes all fetch results associated with a given fetcher.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetch results deleted successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/DeleteManyResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/fetchers/{fetcherId}/fetchResults/{fetchResultId}:
        parameters:
            - $ref: '#/components/parameters/fetcherId'
            - $ref: '#/components/parameters/fetchResultId'
        get:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Get fetch result by fetcher ID and fetch result ID
            operationId: getFetchResultByFetcherIdAndFetchResultId
            description: |
                Retrieves a single fetch result by fetcher ID and fetch result ID.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetch result retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Update fetch result by fetcher ID and fetch result ID
            operationId: updateFetchResultByFetcherIdAndFetchResultId
            description: |
                Partially updates a fetch result by fetcher ID and fetch result ID.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/UpdateFetchResultBody'
            responses:
                '200':
                    description: Fetch result updated successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Delete fetch result by fetcher ID and fetch result ID
            operationId: deleteFetchResultByFetcherIdAndFetchResultId
            description: |
                Deletes a fetch result by fetcher ID and fetch result ID.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetch result deleted successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
components:
    securitySchemes:
        UserBearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: |
                JWT whose `type` claim is `user`. Operations also state the minimum accepted
                `systemRole` (`ADMIN` or `SUPERADMIN`). Role priority is
                `USER < ADMIN < SUPERADMIN`.
        ServiceBearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT whose `type` claim is `service`.

    # ── Parameters ────────────────────────────
    parameters:
        fetcherId:
            name: fetcherId
            in: path
            required: true
            description: |
                Registered fetcher identifier. Current examples:
                `FT_REST_BLUEJAY_REPORTER_LOGS`, `FT_GQL_GITHUB_PROJECTV2_ITEMS`,
                `FT_GQL_GITHUB_PULL_REQUESTS`, `FT_GQL_GITHUB_ISSUES`, and
                `FT_GQL_ZENHUB_ISSUES`.
            schema:
                type: string
                example: FT_GQL_GITHUB_ISSUES
...
        isAsync:
            name: isAsync
            in: query
            required: false
            description: |
                If `true`, a newly claimed external operation continues in the background and the
                response normally contains `IN_PROGRESS`. It has no delaying effect on cache hits or
                on `SNAPSHOT` + `REPLAY`, which can return a terminal result immediately.
            schema:
                type: boolean
                default: false
                example: false

    # ── Request Bodies ────────────────────────
    requestBodies:
...
        GenerateFetchResultBody:
            description: |
                Request body for resolving a persisted fetch result. `CAPTURE` and `REPLAY` express
                the caller's intent; they do not override the fetcher's registered capability.
            required: true
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            temporalContext:
                                $ref: '#/components/schemas/TemporalContext'
                            fetcherConfig:
                                $ref: '#/components/schemas/FetcherConfig'
                        required:
                            - temporalContext
                            - fetcherConfig
                    examples:
                        captureSnapshot:
                            summary: Capture a snapshot at the scheduled instant
                            value:
                                temporalContext:
                                    effectiveAt: '2026-04-10T09:30:00.000Z'
                                    mode: CAPTURE
                                fetcherConfig:
                                    owner: governify
                                    repository: fetcher
                                    token: github_pat_example
                        replaySnapshot:
                            summary: Reuse a previously captured snapshot or return UNAVAILABLE
                            value:
                                temporalContext:
                                    effectiveAt: '2026-04-10T09:30:00.000Z'
                                    mode: REPLAY
                                fetcherConfig:
                                    owner: governify
                                    repository: fetcher
                                    token: github_pat_example
                        captureHistorical:
                            summary: Capture data from a historical-capable source
                            value:
                                temporalContext:
                                    effectiveAt: '2026-04-10T09:30:00.000Z'
                                    mode: CAPTURE
                                fetcherConfig:
                                    latitude: 37.39
                                    longitude: -5.99
                        replayHistorical:
                            summary: Reuse cached data or query a historical source
                            value:
                                temporalContext:
                                    effectiveAt: '2026-04-01T09:30:00.000Z'
                                    mode: REPLAY
                                fetcherConfig:
                                    latitude: 37.39
                                    longitude: -5.99
...
    responses:
        Unauthorized:
            description: Missing, malformed, invalid, expired, or unsupported token
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Authorization header missing or malformed
                        httpStatus: 401
                        appCode: UNAUTHORIZED
                        data: null
                        error:
                            message: Authorization header missing or malformed
        Forbidden:
            description: Authenticated user does not have the required system role
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Insufficient permissions
                        httpStatus: 403
                        appCode: FORBIDDEN
                        data: null
                        error:
                            message: Insufficient permissions
...
    schemas:
...
        SuccessResponse:
            type: object
            description: Standard wrapper for all successful API responses
            additionalProperties: false
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: OK
                httpStatus:
                    type: integer
                    example: 200
                appCode:
                    type: string
                    example: SUCCESS
                data:
                    $ref: '#/components/schemas/JsonValue'
                error:
                    type: object
                    nullable: true
                    example: null
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
...
        FetchResult:
            type: object
            description: |
                Persisted temporal result. Within a fetcher's collection, `effectiveAt` plus `configHash`
                form the unique logical key. `COMPLETED`, `UNAVAILABLE`, and `IN_PROGRESS` results are
                reused. A `FAILED` result is retried by updating the same document.
            additionalProperties: false
            properties:
                _id:
                    type: string
                    description: Stable document identifier, preserved across retries of a `FAILED` result.
                    example: 67c9ff7b2ddf1b3940f9f321
                startDate:
                    type: string
                    format: date-time
                    description: When the current execution was claimed.
                    example: '2026-04-06T10:59:55.000Z'
                endDate:
                    type: string
                    format: date-time
                    nullable: true
                    description: When the current execution reached a terminal state; null while `IN_PROGRESS`.
                    example: '2026-04-06T11:00:00.000Z'
                effectiveAt:
                    type: string
                    format: date-time
                    description: Exact instant represented by this result and part of its logical cache key.
                    example: '2026-04-06T11:00:00.000Z'
                temporalCapability:
                    type: string
                    enum: [SNAPSHOT, HISTORICAL]
                    description: Capability of the fetcher recorded when this result was created or last retried.
                acquisitionMode:
                    type: string
                    enum: [CAPTURE, REPLAY]
                    description: |
                        Mode that created or last retried this document. On a cache hit this can differ
                        from the current request's mode because the stored result is returned unchanged.
                status:
                    type: string
                    enum: [IN_PROGRESS, COMPLETED, UNAVAILABLE, FAILED]
                    description: |
                        Result lifecycle:

                        - `IN_PROGRESS`: an asynchronous or concurrent external operation is running.
                        - `COMPLETED`: data was obtained successfully and is available in `data`.
                        - `UNAVAILABLE`: the requested instant cannot be resolved under the temporal rules;
                          this is a terminal, reusable domain outcome, not an HTTP error.
                        - `FAILED`: the external operation failed; the next generate request reclaims and retries it.
                    example: COMPLETED
                unavailableReason:
                    type: string
                    nullable: true
                    enum: [PAST_SNAPSHOT_NOT_CAPTURED, HISTORICAL_DATA_NOT_AVAILABLE]
                    description: |
                        Reason for `UNAVAILABLE`; null for every other status.

                        - `PAST_SNAPSHOT_NOT_CAPTURED`: replay requested for a snapshot that was not stored.
                        - `HISTORICAL_DATA_NOT_AVAILABLE`: reserved for a historical source that can
                          explicitly report that the instant has no data. Generic source errors produce `FAILED`.
                fetcherConfig:
                    $ref: '#/components/schemas/FetcherConfig'
                configHash:
                    type: string
                    description: SHA-256 hash of the canonical fetcher configuration and part of the logical cache key.
                    pattern: '^[a-f0-9]{64}$'
                    example: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                data:
                    $ref: '#/components/schemas/JsonValue'
                createdAt:
                    type: string
                    format: date-time
                    example: '2026-04-06T10:59:55.000Z'
                updatedAt:
                    type: string
                    format: date-time
                    example: '2026-04-06T11:00:00.000Z'
                __v:
                    type: integer
                    description: MongoDB document version
                    example: 0
            required:
                - _id
                - startDate
                - endDate
                - effectiveAt
                - temporalCapability
                - acquisitionMode
                - status
                - unavailableReason
                - fetcherConfig
                - configHash
                - data
                - createdAt
                - updatedAt
                - __v
...